fix(nextjs): set correct mapping in compute build
- Also improve dir and files structure.
