+++ lib/presets/custom/next/compute/default/handler/routing/index.js
@@ -1,9 +1,7 @@
 // this code is based on cf build next tool (https://github.com/cloudflare/next-on-pages)
 
-// const { cookie, createPCRE } = require('./libs.js');
 const { createPCRE } = require('./libs.js');
 
-// const { parse } = cookie;
 const { RoutesMatcher } = require('./routes-matcher.js');
 const { runOrFetchBuildOutputItem } = require('./utils.js');
