+++ lib/presets/custom/next/compute/default/prebuild/index.js
@@ -3,42 +3,11 @@ import { writeFileSync } from 'fs';
 import { tmpdir } from 'os';
 
 import { VercelUtils, feedback, getAbsoluteLibDirPath } from '#utils';
-import { mapAndAdaptFunctions } from '../mapping/index.js';
-import { assetsPaths } from '../mapping/assets.js';
-import { processVercelOutput } from '../mapping/process-mapping.js';
+import { mapAndAdaptFunctions } from './mapping/index.js';
+import { assetsPaths } from './mapping/assets.js';
+import { processVercelOutput } from './mapping/process-mapping.js';
 
 const { loadVercelConfigs } = VercelUtils;
-
-// TODO: generalize or transform to webpack.
-// Docs: https://esbuild.github.io/plugins/
-const nodeBuiltInModulesPlugin = {
-  name: 'node:built-in:modules',
-  setup(build) {
-    build.onResolve({ filter: /^node:/ }, ({ kind, path }) => {
-      // this plugin converts `require("node:*")` calls, those are the only ones that
-      // need updating (esm imports to "node:*" are totally valid), so here we tag with the
-      // node-buffer namespace only imports that are require calls
-      return kind === 'require-call'
-        ? {
-            path: path.replace('node:', ''),
-            namespace: 'node-built-in-modules',
-          }
-        : undefined;
-    });
-
-    // we convert the imports we tagged with the node-built-in-modules namespace so that instead of `require("node:*")`
-    // they import from `export * from "node:*";`
-    build.onLoad(
-      { filter: /.*/, namespace: 'node-built-in-modules' },
-      ({ path }) => {
-        return {
-          contents: `export * from '${path}'`,
-          loader: 'js',
-        };
-      },
-    );
-  },
-};
 
 /**
  * Construct a record for the build output map.
@@ -60,7 +29,7 @@ function constructBuildOutputRecord(item) {
 
   return `{
 				type: ${JSON.stringify(item.type)},
-				entrypoint: require('${item.entrypoint}')
+				entrypoint: ${item.type === 'node' ? 'null' : `require('${item.entrypoint}')`}
 			}`;
 }
 
@@ -70,11 +39,11 @@ function constructBuildOutputRecord(item) {
  * @param vercelOutput
  */
 function writeOutputReferencesFile(functionsFile, vercelOutput) {
-  console.log('writing references file ...');
+  feedback.prebuild.info('writing references file ...');
   try {
     writeFileSync(
       functionsFile,
-      `export const __BUILD_OUTPUT__ = {${[...vercelOutput.entries()]
+      `globalThis.__BUILD_OUTPUT__ = {${[...vercelOutput.entries()]
         .map(([name, item]) => `"${name}": ${constructBuildOutputRecord(item)}`)
         .join(',')}};`,
     );
@@ -134,13 +103,17 @@ async function run() {
 
   return {
     filesToInject: [
+      // libs used in routing
+      // `${getAbsoluteLibDirPath()}/presets/custom/next/compute/default/handler/routing/libs.js`,
+      // node custom server
+      `${getAbsoluteLibDirPath()}/presets/custom/next/compute/node/handler/index.js`,
+      // file to generate Output. It contains functions references to build
       outputReferencesFilePath,
-      `${getAbsoluteLibDirPath()}/presets/custom/next/compute/default/routing/libs.js`,
     ],
-    globalVars: {
+    workerGlobalVars: {
       __CONFIG__: JSON.stringify(processedVercelOutput.vercelConfig),
     },
-    plugins: [nodeBuiltInModulesPlugin],
+    builderPlugins: [],
   };
 }
