+++ lib/presets/custom/next/compute/default/prebuild/mapping/assets.js
