+++ lib/presets/custom/next/compute/default/prebuild/mapping/index.js
@@ -3,8 +3,9 @@ import { readFileSync, mkdirSync, writeFileSync } from 'fs';
 
 import glob from 'fast-glob';
 
-import { validateFile, normalizePath } from '../../utils/fs.js';
-import { formatRoutePath, stripIndexRoute } from '../../utils/routing.js';
+import { feedback } from '#utils';
+import { validateFile, normalizePath } from '../../../utils/fs.js';
+import { formatRoutePath, stripIndexRoute } from '../../../utils/routing.js';
 import { fixFunctionContent } from '../edge/index.js';
 
 /**
@@ -12,11 +13,25 @@ import { fixFunctionContent } from '../edge/index.js';
  * @param vcConfig
  */
 function isVcConfigValid(vcConfig) {
-  const validRuntime =
-    vcConfig.runtime.match(/edge/) || vcConfig.runtime.match(/node/);
-  const validEntrypoint = vcConfig.entrypoint !== undefined;
+  const isEdge = vcConfig.runtime.match(/edge/) !== null;
+  const isNode = vcConfig.runtime.match(/node/) !== null;
+  const validRuntime = isEdge || isNode;
 
-  return validRuntime && validEntrypoint;
+  // valid framework versions for runtimes
+  const validNextNodeVersions = ['12.3.1'];
+  const validNextEdgeVersions = ['12.3.1', '13.0.0'];
+
+  let validVersion = false;
+  let validEntrypoint = false;
+  if (isEdge) {
+    validEntrypoint = vcConfig.entrypoint !== undefined;
+    validVersion = validNextEdgeVersions.includes(vcConfig.framework.version);
+  } else if (isNode) {
+    validEntrypoint = vcConfig.handler !== undefined;
+    validVersion = validNextNodeVersions.includes(vcConfig.framework.version);
+  }
+
+  return validRuntime && validEntrypoint && validVersion;
 }
 
 /**
@@ -94,19 +109,22 @@ async function mapAndAdaptFunction(
   let isMiddleware = false;
   if (vcObject.content.entrypoint === 'middleware.js' && isEdge) {
     isMiddleware = true;
-    console.log('   Founded middleware!');
+    feedback.prebuild.info('   Founded middleware!');
     // eslint-disable-next-line no-param-reassign
     vcObject.content.entrypoint = 'index.js';
   }
 
-  const codePath = join(path, vcObject.content.entrypoint);
+  const entrypoint = isEdge
+    ? vcObject.content.entrypoint
+    : vcObject.content.handler;
+  const codePath = join(path, entrypoint);
   const relativePath = relative(functionsDir, path);
 
   if (!(await validateFile(codePath))) {
     if (isMiddleware) {
       // We sometimes encounter an uncompiled `middleware.js` with no compiled `index.js` outside of a base path.
       // Outside the base path, it should not be utilised, so it should be safe to ignore the function.
-      console.log(
+      feedback.prebuild.info(
         `Detected an invalid middleware function for ${path}. Skipping...`,
       );
       return {};
@@ -205,7 +223,7 @@ export async function mapAndAdaptFunctions(
     );
   } catch (error) {
     const message = `Error adapting functions: ${error}`;
-    console.log(message);
+    feedback.prebuild.error(message);
     throw Error(message);
   }
