+++ lib/presets/custom/next/compute/default/prebuild/mapping/process-mapping.js
@@ -1,7 +1,8 @@
 import { resolve } from 'path';
 import { rmSync, statSync } from 'fs';
 
-import { addLeadingSlash, stripIndexRoute } from '../../utils/routing.js';
+import { feedback } from '#utils';
+import { addLeadingSlash, stripIndexRoute } from '../../../utils/routing.js';
 
 /**
  *
@@ -70,7 +71,7 @@ function rewriteMiddlewarePaths(processedOutput, middlewarePaths) {
       processedOutput.set(middlewarePath, { ...entry, type: 'middleware' });
       processedOutput.delete(withLeadingSlash);
     } else {
-      console.log(
+      feedback.prebuild.info(
         `Middleware path '${middlewarePath}' does not have a function.`,
       );
     }
