+++ lib/presets/custom/next/compute/handler.js
@@ -1,36 +1,10 @@
-/* eslint-disable import/no-unresolved */
-// .edge files are dynamically generated
-import { assets } from './.edge/next-build/statics.js';
-
+/* eslint-disable */
 import {
   adjustRequestForVercel,
   handleRequest,
-} from './default/routing/index.js';
-
-const createServer =
-  require('./.edge/next-build/custom-server/index.js').default;
-
-globalThis.runNodeCustomServer = async function runNodeCustomServer(request) {
-  const server = await createServer({
-    dir: '.',
-    computeJs: {
-      assets,
-      backends: {
-        httpbin: { url: 'https://httpbin.org/anything/' },
-      },
-    },
-  });
-  const event = {
-    request,
-    client: undefined,
-  };
-  const result = await server.handleFetchEvent(event);
-
-  return result;
-};
+} from 'VULCAN_LIB_PATH/presets/custom/next/compute/default/handler/routing/index.js';
 
 const getStorageAsset = async (request) => {
-  // eslint-disable-next-line no-undef
   const VERSION_ID = AZION_VERSION_ID;
   try {
     const requestPath = new URL(request.url).pathname;
@@ -54,7 +28,7 @@ const getStorageAsset = async (request) => {
  * @param ctx
  */
 async function main(request, env, ctx) {
-  globalThis.process.env = { ...globalThis.process.env, ...env };
+  globalThis.process.env.ASSETS = env.ASSETS;
 
   const adjustedRequest = adjustRequestForVercel(request);
 
@@ -64,9 +38,7 @@ async function main(request, env, ctx) {
       ctx,
       assetsFetcher: env.ASSETS,
     },
-    // eslint-disable-next-line no-undef
     __CONFIG__,
-    // eslint-disable-next-line no-undef
     __BUILD_OUTPUT__,
   );
 }
