+++ lib/presets/custom/next/compute/prebuild.js
@@ -5,7 +5,7 @@ import glob from 'fast-glob';
 import { feedback, getPackageVersion, VercelUtils, Manifest } from '#utils';
 
 import runDefaultBuild from './default/prebuild/index.js';
-// import runNodeBuild from './node/index.js';
+import runNodeBuild from './node/prebuild/index.js';
 
 const { deleteTelemetryFiles, createVercelProjectConfig, runVercelBuild } =
   VercelUtils;
@@ -126,26 +126,24 @@ async function prebuild(buildContext) {
   const nextVersion = getPackageVersion('next');
 
   feedback.prebuild.info('Detected Next.js version:', nextVersion);
-  console.log(' - Build context:', buildContext);
 
   await vercelPrebuildActions();
 
   const projectRuntime = getProjectRuntimes();
 
-  // build routing system and edge functions
-  const prebuildResult = await runDefaultBuild();
-  console.log('\n*** prebuild result:');
-  console.log(prebuildResult);
-
   // build node functions (custom node server)
   if (projectRuntime === 'node' || projectRuntime === 'all') {
-    // await runNodeBuild(nextVersion, buildContext);
-    console.log(' - Node detected!');
+    await runNodeBuild(nextVersion, buildContext);
   }
+
+  // build routing system and edge functions
+  const prebuildResult = await runDefaultBuild();
 
   generateManifest();
 
   feedback.prebuild.success('Next.js build adaptation completed successfully.');
+
+  return prebuildResult;
 }
 
 export default prebuild;
