+++ lib/presets/custom/next/compute/utils/fs.js
@@ -84,7 +84,6 @@ export function validateDir(path) {
 export async function readPathsRecursively(dir) {
   try {
     const files = await readdir(dir);
-
     const paths = await Promise.all(
       files.map(async (file) => {
         const path = resolve(dir, file);
